+++ mint/__main__.py
@@ -1,0 +1,7 @@
+from .app import create_app, main
+
+print("main")
+
+app = create_app()
+
+# main()
